requirements.txt
streamlit
pandas==2.1.4
yfinance
scikit-learn
ta
plotly
newsapi-python
textblob
backtesting
xgboost
lightgbm
pycoingecko
shap==0.44.1
numba==0.58.1
matplotlib
optuna